% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots2.R
\name{boeck_predict}
\alias{boeck_predict}
\title{Plot Observed Frequencies and Model Predictions}
\usage{
boeck_predict(fit_sum = NULL, theta = matrix(0, N, S2), betas = NULL,
  beta_ARS_extreme = NULL, S = NULL, N = 1, J = NULL, revItem = NULL,
  traitItem = NULL, fitModel = NULL, item_order = 1:J)
}
\arguments{
\item{fitModel}{either "ext" or "2012"}
}
\description{
plot expected distribution of frequencies
}

