% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt2irt.R
\docType{data}
\name{boeck_stan_models}
\alias{boeck_stan_models}
\title{Compiled Stan Models (Windows x64)}
\usage{
data(boeck_stan_models)
}
\description{
Compiled Models
}
\details{
\itemize{
  \item boeck_stan_ext
  \item boeck_stan_2012
}
}
\keyword{datasets}

