% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{cat_to_mult}
\alias{cat_to_mult}
\title{Convert to multinomial data.}
\usage{
cat_to_mult(X, C = 5)
}
\arguments{
\item{X}{a N x J   matrix with responses 1...5}
}
\value{
a N x J x C  matrix with response frequencies of 0 and 1
}
\description{
categorical  (responses 1...5) to multinomial (1/0 frequencies) data
}
