% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{create_revItem}
\alias{create_revItem}
\title{Internal function to create revItem from prop.rev}
\usage{
create_revItem(prop.rev = NULL, revItem = NULL, J = NULL,
  n.trait = NULL, traitItem = NULL)
}
\arguments{
\item{prop.rev}{proportion of reversed items (rounded to next integer). Can
be a vector if multiple traits are specified by \code{traitItem}. Only used
if \code{revItem = NULL}.}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=not reversed). Overrides argument \code{prop.rev}.}

\item{J}{number of items}

\item{n.trait}{Integer, number of content-related traits.}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measures, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}
}
\value{
vector \code{revItem} comprised of 0s and 1s.
}
\description{
Creates \code{revItem}  from \code{prop.rev} or simply returns \code{revItem} if supplied.
}
