% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_betas}
\alias{gen_betas}
\title{Generate item parameters.}
\usage{
gen_betas(genModel = c("ext", "2012", "pcm", "steps", "shift", "ext2"),
  J = NULL, betas = NULL)
}
\arguments{
\item{genModel}{Character. The data generating model (either "2012" or "ext").}

\item{J}{number of items. Can be a vector for multiple traits (e.g.,
J=c(10,10,10)).}

\item{betas}{Optional list. May have entries \code{"beta.mrs"},
\code{"beta.ers"}, \code{"beta.trait"}, and/or \code{"beta.ars"}. Each of
those may have arguments passed to \code{\link[truncnorm]{rtruncnorm}}.}
}
\value{
A matrix with \code{J} rows containing the item parameters for MRS,
  ERS, ARS (if \code{genModel != "2012"}), and the target trait.
}
\description{
Function used internally in \code{\link{recovery_irtree}} to generate item parameters.
}
\examples{
J <- 10
# use defaults
betapar <- mpt2irt:::gen_betas("ext", J = J, betas = NULL)
dat <- generate_irtree_ext(N = 20, J = J, betas = betapar, beta_ARS_extreme = 1)

# modify distribution (truncated normal) from which to draw betas, here for MRS
tmp1 <- list("beta.mrs" = list("mean" =  0,
                               "sd"   =  0.3,
                               "a"    = -2,
                               "b"    =  Inf))
betapar <- mpt2irt:::gen_betas("ext", J = J, betas = tmp1)
}
\seealso{
\code{\link[truncnorm]{rtruncnorm}}
}
