% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_irtree_2012}
\alias{generate_irtree_2012}
\title{Generate Data for Original Boeckenholt (2012)}
\usage{
generate_irtree_2012(N, J, betas, traitItem = rep(1, J), theta.vcov = NULL,
  prop.rev = 0.5, cat = T)
}
\arguments{
\item{N}{number of participants}

\item{J}{number of items}

\item{betas}{Jx3 matrix with item properties on three response dimensions (middle, extreme, relevant trait defined by traitItem)}

\item{traitItem}{vector of length J specifying latent traits, e.g., for Big5 from 1...5. standard: unidimensional test, all items measure the same trait.}

\item{theta.vcov}{covariance matrix for middle, extremity, trait(s) (can be a vector of length 3 for uncorrelated processes)}

\item{prop.rev}{proportion of reversed items (rounded to next integer)}

\item{cat}{whether to return categorical data (response categories 1...5) or multinomial data (frequencies of 0 and 1)}
}
\value{
a list containing the generated matrix of responses X, a vector revItem indicating reversed items and true, latent values of the parameters
}
\description{
More general function: \link{generate_irtree_ext}
}
\examples{
# generate data
N <- 20; J <- 10
betas <- cbind(runif(J,0,1), runif(J,0,1), rnorm(J,0,2))
gen <- generate_irtree_2012(N=N, J=J, betas=betas)
}

