% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_irtree_steps}
\alias{generate_irtree_steps}
\title{Generate data for Steps Model.}
\usage{
generate_irtree_steps(N = NULL, J = NULL, revItem = NULL,
  traitItem = NULL)
}
\arguments{
\item{N}{number of persons}

\item{J}{number of items}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=not reversed)}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measures, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}
}
\value{
The function returns a list containing the generated matrix of
  responses X, a vector revItem indicating reversed items and true, latent
  values of the parameters.
}
\description{
Function generates categorical data 1...5 for \code{N} persons and \code{J}
items with paramters from \code{rnorm()}.
}
\section{References}{


Tutz, G. (1997). Sequential models for ordered responses. In W. J. van der
Linden & R. K. Hambleton (Eds.), Handbook of Modern Item Response Theory (pp.
139-152). doi:10.1007/978-1-4757-2691-6_8

Verhelst, N. D., Glas, C. A. W., & de Vries, H. H. (1997). A steps model to 
analyze partial credit. In W. J. van der Linden & R. K. Hambleton (Eds.), 
Handbook of modern item response theory (pp. 123-138).
doi:10.1007/978-1-4757-2691-6_7
}

\examples{
N <- 20
J <- 10
dat <- generate_irtree_steps(N = N, J = J)
}
