% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_dir}
\alias{get_dir}
\title{Get a Directory for Saving Files.}
\usage{
get_dir(dir = NULL)
}
\arguments{
\item{dir}{Character.}
}
\value{
Character, namely, an existing directory.
}
\description{
Check if dir exists and otherwise creates one.
}
