% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{global_OR}
\alias{global_OR}
\title{Discrepency measure for PPC: Global Odds ratio for two items.}
\usage{
global_OR(x, y, threshold = 3, constant = 0.1)
}
\arguments{
\item{x}{Numeric vector containing the responses to the first item.}

\item{y}{Numeric vector containing the responses to the second item.}

\item{threshold}{Numeric vector of length 1. The threshold used for dichotomization.}

\item{constant}{Numeric vector of length 1. A constant added to all four
frequencies in order to avoid dividing by zero.}
}
\value{
The odds ratio.
}
\description{
This function calculates the (natural logarithm of the) global odds ratio of
a pair of items, which are dichotomized beforehand.
}
\references{
Li, T., Xie, C., & Jiao, H. (2017). Assessing fit of alternative
  unidimensional polytomous IRT models using posterior predictive model
  checking. Psychological Methods, 22, 397-408. doi:10.1037/met0000082

Zhu, X., & Stone, C. A. (2012). Bayesian comparison of
  alternative graded response models for performance assessment applications.
  Educational and Psychological Measurement, 72, 774-799.
  doi:10.1177/0013164411434638
}
