% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.R
\name{jacobian_irtree}
\alias{jacobian_irtree}
\title{Check identifiability of models.}
\usage{
jacobian_irtree(N, J, revItem = rep(1, J), traitItem = rep(1, J), type,
  fixed.theta = TRUE, rep = 100, betas, theta)
}
\arguments{
\item{N}{number of persons}

\item{J}{number of items}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=not reversed)}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measures, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}

\item{type}{either "2012", "ext", "ext2", or "ext3"}

\item{fixed.theta}{if TRUE, assumes that mean of theta parameters is assumed to be fixed at zero}

\item{rep}{number of points in parameter space to check identification}

\item{betas}{optional matrix with beta parameters}

\item{theta}{optional matrix with theta parameters (note that the parameters of the first person will be set to to minus the theta-column means if \code{fixed.theta=TRUE})}
}
\description{
Generate random parameter values and test, whether the Jacobian has full rank.
}
\examples{
# Standard model identified, even without reversed items:
jacobian_irtree(N=6, J=8, revItem=rep(1,8), type=2012, rep=5)

}
