% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mcmc.list2stan}
\alias{mcmc.list2stan}
\title{Convert MCMC List to Array}
\usage{
mcmc.list2stan(x)
}
\arguments{
\item{x}{mcmc.list (e.g., from JAGS)}
}
\description{
Converts an mcmc.list (each list entry: rows=MCMC samples; cols=parameters) to a 3 dimensional array (MCMC samples; chains; parameters)
Property of ??? (Internet!)
}

