% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt2irt.R
\docType{package}
\name{mpt2irt}
\alias{mpt2irt}
\alias{mpt2irt-package}
\title{Using MPT models to Get IRT Estimates}
\description{
Combination of Multinomial Processing Tree (MPT) Models with Item Response Theory (IRT) to obtain indivdual estimates for different processes. At the moment, standard models for response styles (choosing middle category, extreme responding, acquiesence; see Boeckenholt, 2012) are implemented in JAGS and Stan.
}
\details{
Requires JAGS (\url{http://mcmc-jags.sourceforge.net}) and/or Stan (\url{http://mc-stan.org}) to sample from the posterior using Markov Chain Monte Carlo (MCMC).
}
\author{
Daniel W. Heck & Hansjoerg Plieninger
}
\references{
Boeckenholt (2012). Modeling Multiple Response Processes in Judgment and Choice. Psychological Methods, 17, 665-678.
}

