% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{mult_to_cat}
\alias{mult_to_cat}
\title{Convert to categorical data.}
\usage{
mult_to_cat(X)
}
\arguments{
\item{X}{a N x J x C matrix with response frequencies of 0 and 1}
}
\value{
N x J   matrix with responses 1...5
}
\description{
multinomial (1/0 frequencies) to categorical data (responses 1...5)
}

