% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots2.R
\name{plot_expected}
\alias{plot_expected}
\title{Plot Expected Frequencies}
\usage{
plot_expected(model, type, X, revItem, traitItem = rep(1, ncol(X)))
}
\arguments{
\item{type}{either "ext" or "2012"}
}
\description{
plot expected distribution of frequencies
}

