% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_Geweke}
\alias{plot_Geweke}
\title{Plot Geweke Statistic}
\usage{
plot_Geweke(fit, parameter = "beta", D = 1, frac1 = 0.1, frac2 = 0.5,
  plot = TRUE, ...)
}
\arguments{
\item{fit}{a fitted object from \link{fit_irtree} (either Boeckenholt 2012 or extended)}

\item{parameter}{Character. The group of parameters for which the stistic is
to be calculated. Using the empty vector (\code{""}) uses all parameters in
\code{fit}.}

\item{D}{Numeric. If the plot should be grouped according to different types
of parameters, \code{D} should be equal to the number of different types of
processes (e.g., \code{D = 4}) for a 1/2/3/ect-dimensional "ext" model or
\code{D=3} for a 1/2/ect-dimensional "Boeckenholt2012" model)}

\item{frac1}{fraction to use from beginning of chain}

\item{frac2}{fraction to use from end of chain}

\item{plot}{Logical. Either return a histogram or the results.}
}
\description{
Plot a Histogram of the Geweke convergence diagnostic.
}
\seealso{
\code{\link{geweke.diag}}
}

