% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_GRS}
\alias{plot_GRS}
\title{Plot Gelman-Rubin statistic.}
\usage{
plot_GRS(fit, parameter = "beta", estimate = "point", plot = TRUE,
  return.odd = TRUE, ...)
}
\arguments{
\item{fit}{a fitted object from \code{\link{fit_irtree}} or preferably from \code{\link{tidyup_irtree_fit}}.
person) or a numeric value specifying the trait value used to plot response
styles (e.g., trait=0 for an average person)}

\item{parameter}{Character. The group of parameters for which the statistic is
to be calculated. Using the empty vector (\code{""}) uses all parameters in
\code{fit}.}

\item{estimate}{Character, either \code{"point"} or \code{"upper"}. Whether
the point estimate or the upper limit of the 95\% CI is to be used.}

\item{plot}{Logical. Either return a histogram or the results.}

\item{return.odd}{Logical. Whether the output for (odd) parameters with a point
estimate > 1.05 should be returned.}

\item{...}{Optional arguments passed to \code{\link[graphics]{hist}}.}
}
\description{
Plot a histogram of the Gelman and Rubin's convergence diagnostic.
}
\seealso{
\code{\link{gelman.diag}}
}
