% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_irtree}
\alias{plot_irtree}
\title{Plot Responses due to Response Styles}
\usage{
plot_irtree(fit, S = 4, N, J, revItem = rep(0, J), traitItem = rep(1, J),
  trait = "sample", return_data = FALSE, tt_names = NULL,
  measure = c("Median", "Mean"), rs_names = NULL, fitMethod = NULL,
  item_order = 1:J)
}
\arguments{
\item{fit}{a fitted object from \link{fit_irtree} (either Boeckenholt 2012 or extended)}

\item{S}{number of distinct MPT-parameters (2012: 3; extended: 4)}

\item{N}{number of participants}

\item{J}{number of items}

\item{revItem}{a vector with 1=reversed / 0=not reversed}

\item{traitItem}{a vector specifying the trait of the item (e.g., with values from 1 to 5 for Big5)}

\item{trait}{either "sample" (using the posterior mean estimates for each person) or a numeric value specifying the trait value used to plot response styles (e.g., trait=0 for an average person)}

\item{return_data}{Logical indicating whether the data frame used for plotting should be returned.}

\item{tt_names}{Optional character vector with the name(s) of the target trait(s).}

\item{measure}{Character vector that indicates whether the mean (default) or the median of the posterior distribution should be plotted.}
}
\description{
Plot proportion of responses due to RS
}

