% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots2.R
\name{plot_responses}
\alias{plot_responses}
\title{Plot Observed Frequencies}
\usage{
plot_responses(X, revItem, traitItem, points = 5)
}
\arguments{
\item{X}{matrix with observed responses for all participnats (by row) and items (by column)}

\item{revItem}{use different colors, if specified (as a vector with 1=reversed, 0=not reversed)}

\item{traitItem}{use as labels for separate traits}

\item{points}{how many resposne categories in Likert scale}
}
\description{
Plot histograms of observed response frequencies separately for each item. This is especially helpful to judge whether simulated data are reasonable (i.e., unimodal etc.).
}

