% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_singlefit}
\alias{plot_singlefit}
\title{Plot True and Estimated Parameters}
\usage{
plot_singlefit(N, J, S, model, jags.samp, stan.samp, true.par,
  traitItem = rep(1, J), revItem = rep(1, J))
}
\arguments{
\item{N}{number of participants}

\item{J}{number of items}

\item{S}{number of latent processes to be measured}

\item{model}{either "2012" or "ext"}

\item{jags.samp}{a fitted mcmc.list from JAGS}

\item{stan.samp}{a fitted Stan array}

\item{true.par}{a list with true values for beta and theta}

\item{traitItem}{if more than a single trait is measured, items measuring different traits are shown in different colors}

\item{revItem}{if specified, reversed items are shown as different points}

\item{number}{of theta parameters to plot (either 3 or 4)}
}
\description{
Plot estimates from a single fit using jags and stan (or estimates and true values). Only two object of jags.samp/stan.samp/true.par are allowed.
}
\examples{
\dontrun{
# generate data (Boeckenholt, 2012)
N <- 40; J <- 30; n.trait <- 3
traitItem=rep(1:n.trait, each=J/n.trait)
betas <- cbind(runif(J,0,1),
              runif(J,0,1),
              rnorm(J,0,2))
              
# analyze and plot
gen <- generate_irtree_2012(N=N, J=J, betas=betas, traitItem=traitItem, prop.rev=.5)
fit <- fit_irtree(gen$X, gen$revItem, gen$traitItem, fitModel="2012",
                 M= 500, n.chains=4, thin=1, fitMethod="jags", format="jags")
plot_singlefit(N, J, 2+n.trait, "2012", jags.samp=fit, revItem=gen$revItem,traitItem=gen$traitItem,
               true.par=list(theta=gen$theta, betas=gen$betas))
               
# check with stan
fit2 <- fit_irtree(gen$X, gen$revItem, gen$traitItem, fitModel="2012",
                  M= 500, n.chains=4, thin=1, fitMethod="stan", format="jags")
plot_singlefit(N, J, 2+n.trait, "2012", jags.samp=fit, stan=fit2, 
               revItem=gen$revItem,traitItem=gen$traitItem)
}
}

