% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{post_prob_irtree}
\alias{post_prob_irtree}
\title{Calculate posterior predictive probabilities.}
\usage{
post_prob_irtree(fit_sum = NULL, mcmc.objects = NULL, iter = 100,
  N = NULL, traitItem = NULL, revItem = NULL, fitModel = NULL,
  new_theta = FALSE)
}
\arguments{
\item{fit_sum}{List. Output from \code{\link{summarize_irtree_fit}} that
contains \code{mcmc.objects}.}

\item{mcmc.objects}{a list of MCMC or runjags objects, all with the same number of chains and matching variable names, or a single MCMC object/list or runjags object.  No default.}

\item{iter}{Numeric. Number of iterations to use, the maximum is the total
number of retained iterations (via \code{\link{fit_irtree}}).}

\item{N}{Numeric. Number of persons for whom posterior predictives should be
drawn. Should be equal to the number of persons in the sample if
\code{new_theta = FALSE}.}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measured, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=regular)}

\item{fitModel}{Character. Either \code{"2012"} (Boeckenholt Model without
acquiescence) or \code{"ext"} (Acquiescence Model). Details about all
implemented models are described in the section Models below.}

\item{new_theta}{Logical. Wheter to calculate the probabilities for the
persons in the sample or for out-of-sample predictions for \code{N} 'new'
persons (\code{new_theta = TRUE}).}
}
\value{
Returns an array of probabilities of dimension \code{iter} x \code{N}
  x J x 5 (for J items with 5 response categories).
}
\description{
Function takes an MCMC list of posterior samples and calculates the
model-predicted probabilities. This can either be done for the persons in the
sample or for out-of-sample predictions for new persons (\code{new_theta =
TRUE}).
}
\examples{
\dontrun{
J <- 10
betas <- cbind(rnorm(J, .5), rnorm(J, .5), rnorm(J, 1.5), rnorm(J, 0))
dat <- generate_irtree_ext(N = 20, J = J, betas = betas, beta_ARS_extreme = .5)

# fit model
res1 <- fit_irtree(dat$X, revItem = dat$revItem, M = 200)
res2 <- summarize_irtree_fit(res1)

# posterior predictive probabilities
res3 <- post_prob_irtree(res2)
dim(res3)
}
}
