% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{posterior_expected}
\alias{posterior_expected}
\title{Calculate expected responses on the basis of posterior predicitive
probabilities.}
\usage{
posterior_expected(prob = NULL)
}
\arguments{
\item{prob}{Numeric array of dimension R x N x J x 5 (for N persons, J items
with 5 categories, and R iterations).}
}
\value{
Array of expected responses of dimension R x N x J (for N persons, J
  items, and R iterations).
}
\description{
The function takes the posterior predicitve probabilities (from
\code{\link{post_prob_irtree}}) and calculates for every
iteration-person-item combination the expected response. This is needed for
Yen's Q3 (see \code{\link{statistic_Q3}}).
}
