% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{pp_irtree}
\alias{pp_irtree}
\title{Draw posterior predictive values.}
\usage{
pp_irtree(fit_sum = NULL, mcmc.objects = NULL, iter = 100, probs = NULL,
  N = NULL, traitItem = NULL, revItem = NULL, fitModel = NULL)
}
\arguments{
\item{fit_sum}{List. Output from either \code{\link{tidyup_irtree_fit}} or
\code{\link{summarize_irtree_fit}} that contains \code{mcmc.objects}.}

\item{mcmc.objects}{a list of MCMC or runjags objects, all with the same number of chains and matching variable names, or a single MCMC object/list or runjags object.  No default.}

\item{iter}{Numeric. Number of iterations per chain to use for generating
posterior predictives.}

\item{probs}{Numeric. Vector of probabilities (passed to
\code{\link[stats]{quantile}}) that is used to calculate quantiles of the
posterior predictive distribution. Argument may be named (used to generate
colnames). Defaults to \code{c(.025, .975, .16, .84)}.}

\item{N}{Numeric. Number of persons for whom posterior predictives should be
drawn.}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measures, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=not reversed)}

\item{fitModel}{Character. Either \code{"2012"} (Boeckenholt Model without 
acquiescence), or \code{"ext"} (Acquiescence Model), or \code{"pcm"}
(partial credit model), or \code{"steps"} (Steps Model [Verhelst; Tutz]),
or \code{"shift"} (shift model, i.e., \code{"2012"} +  ars-shift).}
}
\value{
Returns a data frame containing the specified quantiles of the
  posterior predictive distribution for every category of every item.
}
\description{
Function takes an MCMC list and draws posterior predictive values.
}
\details{
Given an MCMC list, this function draws theta values for \code{N} persons
from the "current" multivariate normal distribution and---given the current
item parameters---calculates the probabilities (given the fitted model) for
every item--person combination. Posterior predictive responses (i..e, an
integer between 1 and 5) are then drawn based on these probabilities and this
is done \code{iter} times per chain. In each iteration and for every item,
the responses are aggregated across persons resulting in a frequency table
(i.e., a distribution), and---across iterations---this finally results in a
posterior predictive distribution for every category for every item. This
distribution is summarized with the quantiles given in \code{probs} and a
data frame containing these quantiles for every item is returned.
}
\examples{
\dontrun{
J <- 10
betas <- cbind(rnorm(J, .5), rnorm(J, .5), rnorm(J, 1.5), rnorm(J, 0))
dat <- generate_irtree_ext(N = 20, J = J, betas = betas, beta_ARS_extreme = .5)

# fit model
res1 <- fit_irtree(dat$X, revItem = dat$revItem, M = 200)
res2 <- summarize_irtree_fit(res1)

# posterior predictives for 512 hypothetical persons
res3 <- pp_irtree(res2, iter = 10, N = 512)
}
}
