% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{ppc_irtree}
\alias{ppc_irtree}
\title{Posterior predicitve checks and p-values for several discrepency measures.}
\usage{
ppc_irtree(prob = NULL, statistics = c("item_cor", "OR", "Q3", "ISD",
  "resp"), fit = NULL, X = NULL, revItem = NULL, traitItem = NULL)
}
\arguments{
\item{prob}{Numeric array of dimension R x N x J x 5 (for N persons, J items
with 5 categories, and R iterations).}

\item{statistics}{Character vector of length >= 1 specifying the discrepency
measures to use.}

\item{fit}{a fitted object from \code{\link{fit_irtree}} or preferably from
\code{\link{summarize_irtree_fit}}.}

\item{X}{Numeric matrix of dimension N x J containing the observed item responses.

\itemize{
\item \code{item_cor}: Item-total correlation, i.e., polyserial correlation between response and total score.
\item \code{OR}: Odds ratio for (dichotomized) pairs of items.
\item \code{Q3}: Yen's Q3 statistic.
\item \code{ISD}: Item-score-distribution, i.e., Pearson's X2 of residuals.
\item \code{resp}: Observed responses (integer), predicted responses (integer), and expected responses (numeric).
}}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=regular)}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measured, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}
}
\value{
Returns an object of class \code{ppc}
}
\description{
This function takes posterior predictive probabilities (from
\code{\link{post_prob_irtree}}), and compares---for several discrepency
measures---observations and predictions both descriptively and via posterior
predicitive p-values.
}
\examples{
\dontrun{
J <- 10
betas <- cbind(rnorm(J, .5), rnorm(J, .5), rnorm(J, 1.5), rnorm(J, 0))
dat <- generate_irtree_ext(N = 20, J = J, betas = betas, beta_ARS_extreme = .5)

# fit model
res1 <- fit_irtree(dat$X, revItem = dat$revItem, M = 200)
res2 <- summarize_irtree_fit(res1)

# posterior predictive checking
res3 <- post_prob_irtree(res2)
res4 <- ppc_irtree(prob = res3, fit = res1)
res4
}

}
\references{
Levy, R. (2011). Posterior predictive model checking for
  conjunctive multidimensionality in item response theory. Journal of
  Educational and Behavioral Statistics, 36, 672-694.
  doi:10.3102/1076998611410213

Li, T., Xie, C., & Jiao, H. (2017). Assessing fit of alternative
  unidimensional polytomous IRT models using posterior predictive model
  checking. Psychological Methods, 22, 397-408. doi:10.1037/met0000082

Sinharay, S., Johnson, M. S., & Stern, H. S. (2006). Posterior
  predictive assessment of item response theory models. Applied Psychological
  Measurement, 30, 298-321. doi:10.1177/0146621605285517

Zhu, X., & Stone, C. A. (2012). Bayesian comparison of
  alternative graded response models for performance assessment applications.
  Educational and Psychological Measurement, 72, 774-799.
  doi:10.1177/0013164411434638
}
\seealso{
\code{\link{print.ppc}} for summarizing the results and
  \code{\link{ppc_resp_irtree}} for summarizing posterior predictive response
  frequencies.
}
