% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{print.ppc}
\alias{print.ppc}
\title{Print posterior predictive p-values}
\usage{
\method{print}{ppc}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{List as returned from \code{\link{ppc_irtree}}}

\item{na.rm}{Logical argument passed to \code{\link[base]{mean}}
\code{\link[stats]{median}}. If \code{na.rm = FALSE}, the present print
method will return \code{NA} if at least one PPP-value could not be
calculated for a discrepency measure. Thus this is extremely helpful for
identifying problems; set it to \code{TRUE} only if you know what you are
doing.}

\item{...}{Further arguments passed to \code{\link{print.table}}.}
}
\value{
Table of PPP-values
}
\description{
This function calculates and prints PPP-values for objects of class
\code{ppc}. More precisely, it prints---across all items---the average of the
PPP-values as well as the percentage of extreme PPP-values. For example, for
a data set comprised of 20 items, 190 odds ratios can be calculated, namely,
one for each pair of items. The present print method calculates the avarage
PPP-value (should probably be close to 0.5) as well as the percentage of
those 190 PPP-values that are extreme (should hopefully be close to 0).
}
