% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pp_man.R
\name{sample_pp}
\alias{sample_pp}
\title{Sample Posterior Predictive Responses}
\arguments{
\item{prob}{Numeric matrix with K columns}
}
\value{
Numeric vector of length \code{nrow(prob)} containing integerish
  responses between 1 and K
}
\description{
This function samples posterior predictive responses (x=1,...,K) given a
matrix of probabilities. It is implemented in C++ to increase speed.
}
\examples{
prob <- matrix(runif(5*20), 20)
mpt2irt:::sample_pp(prob)
}
