% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stan2mcmc.list}
\alias{stan2mcmc.list}
\title{Stan to coda object}
\usage{
stan2mcmc.list(fit)
}
\arguments{
\item{fit}{fitted Stan object}
}
\description{
Convert a 3 dimensional array (MCMC samples; chains; parameters) to an mcmc.list object for analysis with the coda package
}
\keyword{internal}
