% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{statistic_item_cor}
\alias{statistic_item_cor}
\title{Discrepency measure for PPC: Item-total correlation.}
\usage{
statistic_item_cor(resp = NULL, revItem = NULL, traitItem = NULL)
}
\arguments{
\item{resp}{Numeric matrix of dimension N x J (for N persons and J items)
containing the observed item responses.}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=regular)}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measured, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}
}
\value{
Vector of length J of polyserial correlations.
}
\description{
This function calculates the polyserial correlation between a person's total
score and his or her item response. This discrepency measure is useful to
detect misfit due to a missing item-slope/-discrimination parameter. It may
not be suitable for the response style models discussed herein, because these
models assume that the total score as well as the item response is a
composite of target trait and response styles.
}
\references{
Li, T., Xie, C., & Jiao, H. (2017). Assessing fit of alternative
  unidimensional polytomous IRT models using posterior predictive model
  checking. Psychological Methods, 22, 397-408. doi:10.1037/met0000082

Zhu, X., & Stone, C. A. (2012). Bayesian comparison of
  alternative graded response models for performance assessment applications.
  Educational and Psychological Measurement, 72, 774-799.
  doi:10.1177/0013164411434638
}
