% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{statistic_item_score}
\alias{statistic_item_score}
\title{Discrepency measure for PPC: Item score distribution X2.}
\usage{
statistic_item_score(resp = NULL, prob = NULL)
}
\arguments{
\item{resp}{Numeric matrix of dimension N x J (for N persons and J items)
containing the responses.}

\item{prob}{Numeric array of dimension R x N x J x 5 (for N persons, J items
with 5 categories, and R iterations).}
}
\value{
Vector of length J containing X2 values.
}
\description{
This function calculates the item score distribution, i.e., the discrepancy
between observed frequencies and posterior predictive frequencies for each
item using Pearson's X2 statistic.
}
\references{
Zhu, X., & Stone, C. A. (2012). Bayesian comparison of
  alternative graded response models for performance assessment applications.
  Educational and Psychological Measurement, 72, 774-799.
  doi:10.1177/0013164411434638
}
