% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{statistic_OR}
\alias{statistic_OR}
\title{Discrepency measure for PPC: Global Odds ratio for three or more items.}
\usage{
statistic_OR(resp)
}
\arguments{
\item{resp}{Numeric matrix of dimension N x J (for N persons and J items)
containing the responses.}
}
\value{
Vector of length 'J choose 2' odds ratios.
}
\description{
This function calculates the (natural logarithm of the) global odds ratio of
pairs of items, which are dichotomized beforehand.
}
\references{
Li, T., Xie, C., & Jiao, H. (2017). Assessing fit of alternative
  unidimensional polytomous IRT models using posterior predictive model
  checking. Psychological Methods, 22, 397-408. doi:10.1037/met0000082

Zhu, X., & Stone, C. A. (2012). Bayesian comparison of
  alternative graded response models for performance assessment applications.
  Educational and Psychological Measurement, 72, 774-799.
  doi:10.1177/0013164411434638
}
