% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred.R
\name{statistic_Q3}
\alias{statistic_Q3}
\title{Discrepency measure for PPC: Yen's Q3.}
\usage{
statistic_Q3(resp = NULL, exp = NULL)
}
\arguments{
\item{resp}{Numeric matrix of dimension N x J (for N persons and J items)
containing the responses.}

\item{exp}{Numeric matrix of dimension N x J (for N persons and J items)
containing the expected responses.}
}
\value{
Vector of length J of polyserial correlations.
}
\description{
This function calculates Yen's Q3 statistic.
}
\references{
Li, T., Xie, C., & Jiao, H. (2017). Assessing fit of alternative
  unidimensional polytomous IRT models using posterior predictive model
  checking. Psychological Methods, 22, 397-408. doi:10.1037/met0000082

Yen, W. M. (1993). Scaling performance assessments: Strategies
  for managing local item dependence. Journal of Educational Measurement, 30,
  187-213. doi:10.1111/j.1745-3984.1993.tb00423.x

Zhu, X., & Stone, C. A. (2012). Bayesian comparison of
  alternative graded response models for performance assessment applications.
  Educational and Psychological Measurement, 72, 774-799.
  doi:10.1177/0013164411434638
}
