% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fit.R
\name{summarize_irtree_fit}
\alias{summarize_irtree_fit}
\title{Calculate summary for a fitted model.}
\usage{
summarize_irtree_fit(fit, fitMethod = NULL, interact = FALSE, ...)
}
\arguments{
\item{fit}{a fitted object from \code{\link{fit_irtree}}.}

\item{fitMethod}{whether to use JAGS or Stan}

\item{interact}{logical. If set to \code{TRUE}, the function may be prompt for input.}

\item{...}{Further arguments passed to \code{\link[coda]{summary.mcmc.list}}.}
}
\value{
Returns a list containing the input \code{fit} as well as an MCMC list and \code{summary}.
}
\description{
Function takes a fitted model returned from \code{\link{fit_irtree}} and
calculates summaries for all parameters using coda's
\code{\link[coda]{summary.mcmc.list}}.
}
\details{
The difference between the present function and directly calling coda's
\code{\link[coda]{summary.mcmc.list}} is that, herein, the correlations of the thetas
are summarized in addition to the covariances.
}
\examples{
\dontrun{
# generate data
N <- 20
J <- 10
betas <- cbind(rnorm(J, .5), rnorm(J, .5), rnorm(J, 1.5), rnorm(J, 0))
dat <- generate_irtree_ext(N = N, J = J, betas = betas, beta_ARS_extreme = .5)

# fit model
res1 <- fit_irtree(dat$X, revItem = dat$revItem, M = 200)
res2 <- summarize_irtree_fit(res1)
head(res2$summary$statistics)
head(res2$summary$quantiles)
}
}
