% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fit.R
\name{tidyup_irtree_fit}
\alias{tidyup_irtree_fit}
\title{Summarize and Tidy up a Fitted Model}
\usage{
tidyup_irtree_fit(fit, S = NULL, N = NULL, J = NULL, revItem = NULL,
  traitItem = NULL, fitMethod = NULL, fitModel = NULL,
  measure = c("Median", "Mean"), tt_names = NULL, plot = TRUE, ...)
}
\arguments{
\item{fit}{a fitted object from \code{\link{fit_irtree}} or preferably from \code{\link{summarize_irtree_fit}}.}

\item{S}{number of distinct MPT-parameters (2012: 3; extended: 4)}

\item{N}{number of participants}

\item{J}{number of items}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=not reversed)}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measures, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}

\item{fitMethod}{whether to use JAGS or Stan}

\item{fitModel}{Character. Either \code{"2012"} (without acquiescence) or \code{"ext"}
(assumes that in case of ARS, the probability of choosing the extreme
category 5 is determined by ERS) or \code{"pcm"} (partial credit model).}

\item{measure}{Character vector that indicates whether the mean (default) or the median of the posterior distribution should be plotted.}

\item{tt_names}{Optional character vector with the name(s) of the target trait(s).}

\item{plot}{Logical. Whether a plot should be produced using \code{\link{plot_irtree}} and returned.}

\item{...}{Further arguments passed to \code{\link{plot_irtree}}.}
}
\description{
Function takes a fitted model returned from \code{\link{fit_irtree}} and
returns a tidy summary of the fitted parameters.
}

