% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_irtree}
\alias{fit_irtree}
\title{Analyze with Boeckenholt Model for Response Styles}
\usage{
fit_irtree(X, revItem = NULL, traitItem = rep(1, ncol(X)), df = NULL,
  V = NULL, fitModel = c("ext", "2012", "pcm"), fitMethod = c("stan",
  "jags"), outFormat = NULL, startSmall = TRUE, M = 1000, warmup = 1000,
  n.chains = 2, thin = 1, method = "parallel", return_defaults = TRUE,
  add2varlist = NULL, cores = NULL, summarise = FALSE, N2 = 2, ...)
}
\arguments{
\item{X}{an N x J matrix of observed responses for categories 1...5 (use
\code{\link{mult2cat}} to transform a multinomial frequency matrix with 1s/0s to
responses from 1...5)}

\item{revItem}{vector of length J specifying reversed items (1=reversed,
0=not reversed)}

\item{traitItem}{vector of length J specifying the underlying traits (e.g.,
indexed from 1...5). Standard: only a single trait is measured by all
items. If the Big5 are measures, might be something like
c(1,1,1,2,2,2,...,5,5,5,5)}

\item{df}{degrees of freedom for wishart prior on covariance of traits
(standard/minimum: number of processes + 1)}

\item{V}{prior for wishart distribution (standard: diagonal matrix)}

\item{fitModel}{Character. Either \code{"2012"} (without acquiescence) or \code{"ext"}
(assumes that in case of ARS, the probability of choosing the extreme
category 5 is determined by ERS) or \code{"pcm"} (partial credit model).}

\item{fitMethod}{whether to use JAGS or Stan}

\item{outFormat}{either "mcmc.list" (can be analyzed with coda package) or
"stan" or "runjags"}

\item{startSmall}{Whether to use random starting values for
beta sampled from "wide" (startSmall=F, generated within JAGS) or "narrow"
priors (startSmall=T; beta and theta closer to 0; might solve problems with
slow convergence of some chains for extreme starting values).}

\item{M}{number of MCMC samples (after warmup)}

\item{warmup}{number of samples for warmup (in JAGS: 3/4 for adaption, 1/4
for burnin)}

\item{n.chains}{number of MCMC chains (and number of CPUs used)}

\item{thin}{thinning of MCMC samples}

\item{add2varlist}{Additional variables to monitor (e.g., \code{c("deviance",
"pd", "popt", "dic")} for JAGS)}

\item{N2}{Numeric. Number of persons for whom to draw posterior predictives.
Specify equal to \code{nrow(X)} in order to draw values for all persons.
This is mainly implemented for efficiency reasons in order to avoid
massivly drawing samples which the user is not interested in.}

\item{...}{further arguments passed to \code{\link[rstan]{sampling}} (for Stan) or \code{\link[runjags]{run.jags}} (for JAGS)}
}
\description{
Using either the standard 2012 model (only response styles towards middle + extreme responding) or the extended version (additionally, aquiescence).
}
\details{
The estimated parameters are ordered as follows:
\itemize{
\item Model "2012": S=2+number of traits
\itemize{
 \item theta[i,1:S] = c(middle, extreme, trait(s))
 \item betas[j,1:3] = c(middle, extreme, trait) (which trait depends on traitItem!)
 }
\item Model "ext": S=3+number of traits
\itemize{
 \item theta[i,1:S] = c(middle, extreme, acquiesence, trait(s))
 \item betas[j,1:4] = c(middle, extreme, acquiesence, trait) (which trait depends on traitItem!)
 }
}
If more than a single trait is measured, theta has more columns accordingly (e.g., theta[i,1:6]=c(mid, extr, acq, trait1,..., trait3))

 Note that the progress of Stan is shown in a text file in the
  working directory ("_Stanprogress.txt")
}
\examples{
\dontrun{
}
}

