% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_betas}
\alias{gen_betas}
\title{Generate Item Parameters}
\usage{
gen_betas(genModel = genModel, J = J, betas = betas)
}
\arguments{
\item{genModel}{Character. The data generating model (either "2012" or "ext").}

\item{J}{number of items. Can be a vector for multiple traits (e.g.,
J=c(10,10,10)).}

\item{betas}{Optional list. May have entries \code{"beta.mrs"},
\code{"beta.ers"}, \code{"beta.trait"}, and/or \code{"beta.ars"}. Each of
those may have arguments passed to \code{\link[truncnorm]{rtruncnorm}}.}
}
\value{
A matrix with \code{J} rows containing the item parameters for MRS, ERS,---if \code{genModel != "2012"}---ARS, and trait.
}
\description{
Function used internally in \code{\link{recovery_irtree}} to generate item parameters.
}
\seealso{
\code{\link[truncnorm]{rtruncnorm}}
}

