% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.R
\name{jacobian_irtree}
\alias{jacobian_irtree}
\title{Check Identifiability of Models}
\usage{
jacobian_irtree(N, J, revItem = rep(1, J), traitItem = rep(1, J), type,
  fixed.theta = TRUE, rep = 100, betas, theta)
}
\arguments{
\item{N}{number of participants}

\item{J}{number of items}

\item{revItem}{vector with 0/1 for normal/reversed items}

\item{traitItem}{vector of length J specifying latent traits, e.g., for Big5 from 1...5 . standard: unidimensional test, all items measure the same trait.}

\item{type}{either "2012", "ext", "ext2", or "ext3"}

\item{fixed.theta}{if TRUE, assumes that mean of theta parameters is assumed to be fixed at zero}

\item{rep}{number of points in parameter space to check identification}

\item{betas}{optional matrix with beta parameters}

\item{theta}{optional matrix with theta parameters (note that the parameters of the first person will be set to to minus the theta-column means if \code{fixed.theta=TRUE})}
}
\description{
Generate random parameter values and test, whether the Jacobian has full rank.
}
\examples{
# Standard model identified, even without reversed items:
jacobian_irtree(N=6, J=8, revItem=rep(1,6), type=2012, rep=5)

# Extended model NOT identified without reversed items:
jacobian_irtree(N=6, J=8, revItem=rep(1,8), type="ext", rep=5)

# Ext3 identified with reversed items and 2 traits:
jacobian_irtree(N=6, J=8, revItem=rep(0:1,4), 
               traitItem=rep(1:2,each=4), type="ext3", rep=5)
}
